+++ CHANGELOG.md
@@ -9,6 +9,8 @@ on every vscode update a new version will be release
 
 * you would have to check the CHANGELOG for any breaking, (major), minor or patched updates which will be denoted respectively
 
+
+
 ## [1.85.1000] - 2023-12-27
 * Extension made available to the public ready for use
 
@@ -17,3 +19,7 @@ on every vscode update a new version will be release
 * [UPDATE] Set Default Option Available and persists in settings accordingly
 * [FIX] fixed an issue where auto open would trigger on unrelated files
 * [UPDATE] ensured output channel is cleared on every call to openRelatedFiles
+
+## [1.85.1002] - 2023-12-27
+* [UPDATE] Can toggle whether you want to reset the layout on every new file that gets openend
+* [CHANGE] renamed command from Windmillcode Open Related Files: Auto Open to Windmillcode Open Related Files: Toggle Auto Open